feat(rpg): Фаза 3 шаг 4а — int-транспорт урона от броска до двери; combat.md честный
Флоаты боевых законов добиты на всём пути раны: apply_damage/DamageResult/
apply_spell_damage/лог-колбэки = int, Projectile.damage = int32 (бросается
на касте/выстреле), chainDecay стал целым процентом chainDecayPct (0.70→70)
и цепная математика int, падение округляется ОДИН раз на пороге двери
(кинетика остаётся float — это физика, не закон). Казнь дробного бара =
ceil-удар (оверкилл меньше поинта, выживших нет).

Health остаётся float-ХРАНИЛИЩЕМ с целыми писателями — свип хранилища
(и бамп макро-снимка, который он тянет) отложен отдельным заходом, зафиксирован
в combat.md «Left standing».

combat.md переписан под осевшие законы: кубы/9×9/гибрид/крит, новые дома
(damage_types.h), честные bias-оговорки авторезолва, живой список
недоделанного (носимая броня не защищает под землёй — компонент не
проецируется на тело).

ctest 94/94; смоуки melee / cast_spell / missile_feedback / reputation_hit
1,7,999 PASS.

+++ src/sub/damage.h
@@ -95,15 +95,18 @@ struct DamageSource {
 };
 
 struct DamageResult {
-    float applied = 0.0f;  // post-mitigation hp actually subtracted; 0 = no-op
-    bool lethal = false;   // this blow drove hp to zero
+    int applied = 0;      // post-mitigation hp actually subtracted; 0 = no-op
+    bool lethal = false;  // this blow drove hp to zero
 };
 
 // `type` is WHICH of the nine columns the blow argues with (CANON S13: the
 // 9×9 symmetry) — it selects the armour column inside the door and nothing
 // else; a kind whose row says armour is not in the way ignores it entirely.
+// `amount` is integer like every combat quantity (phase 3: combat floats are
+// dead; the physics that PRODUCED a number — a fall's kinetics — stays float
+// and rounds once, at this threshold).
 DamageResult apply_damage(entt::registry& reg, entt::entity target,
-                          const DamageSource& src, float amount,
+                          const DamageSource& src, int amount,
                           DamageKind kind, DamageType type, EventBus* bus);
 
 // An execution: one blow of exactly the target's remaining hp, through the
